updating titles print
Former-commit-id: 66dea8fcfa0f3d766aab147dbb03cf2df8dd5e31

+++ java/src/xmipp/ij/commons/ImagePlusReader.java
@@ -21,7 +21,7 @@ public abstract class ImagePlusReader {
     
     protected long modified;
     
-    protected int index = -1;
+    protected long index = -1;
     protected boolean normalize;
     protected double normalize_min;
     protected double normalize_max;
